requirements.txt
celery==5.3.1
cryptography==41.0.1
dependency-injector==4.41.0
django==3.2.16
django-environ==0.9.0
django-minio-storage==0.5.6
django-redis==5.2.0
django-select2==8.1.1
django-simple-menu==2.1.1
django-tree-queries==0.14.0
django-widget-tweaks==1.4.12
psycopg2==2.9.5
pydantic==2.7.0
requests==2.30.0
xmltodict==0.13.0
zeep==4.2.1
